+++ Tests/SentryTests/Integrations/SentryNetworkTrackerTests.swift
@@ -155,6 +155,17 @@ class SentryNetworkTrackerTests: XCTestCase {
         XCTAssertEqual(spans.count, 1)
     }
     
+    func testIntercepted() {
+        let sut = fixture.getSut()
+        let task = createInterceptedRequest()
+        sut.urlSessionTaskResume(task)
+        
+        let tracker = fixture.tracker
+        let spans = getStack(tracker: tracker)
+        
+        XCTAssertEqual(spans.count, 0)
+    }
+    
     func testCaptureRequestDuration() {
         let sut = fixture.getSut()
         let task = createDataTask()
@@ -305,5 +316,12 @@ class SentryNetworkTrackerTests: XCTestCase {
         var request = URLRequest(url: SentryNetworkTrackerTests.testURL)
         request.httpMethod = method
         return URLSessionStreamTaskMock(request: request)
+    }
+    
+    private func createInterceptedRequest() -> URLSessionDownloadTaskMock {
+        let request = NSMutableURLRequest(url: SentryNetworkTrackerTests.testURL)
+        URLProtocol.setProperty(true, forKey: SENTRY_INTERCEPTED_REQUEST, in: request)
+        request.httpMethod = "GET"
+        return URLSessionDownloadTaskMock(request: request as URLRequest)
     }
 }